bbg_client/requirements.txt
ansicolors==1.0.2
appdirs==1.4.3
Babel==2.4.0
closure==20161201
configparser==3.5.0
decorator==4.0.11
defusedxml==0.5.0
Django==1.9.13
django-appconf==1.0.2
django-compressor==2.1.1
django-compressor-toolkit==0.6.0
django-filter==1.0.2
django-gulp==2.5.1
django-redis==4.7.0
djangorestframework==3.6.2
djangorestframework-jsonapi==2.1.1
env-tools==2.0.0
futures==3.0.5
httplib2==0.10.3
inflection==0.3.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
lxml==3.7.3
Markdown==2.6.8
oauth2==1.9.0.post1
oauthlib==2.0.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
ply==3.10
prompt-toolkit==1.0.14
psutil==4.3.0
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
slimit==0.8.1
social-auth-app-django==1.1.0
social-auth-core==1.2.0
tini==3.0.1
traitlets==4.3.2
wcwidth==0.1.7
